Update structure of the repo (#53)
* Rearranged initial directory structure

* update directory
