--- References_and_Guides/GenanceSim_Commands.md
@@ -1,43 +1,0 @@
-# GenAnce Commands
-
-```-b```
-
-- Simulation will stop when all degnomes are identical.
-
-```-c chromosome_length```
-
-- Set chromosome length for the current simulation.
-- Default chromosome length is 10.
-
-```-g num_generations```
-
-- Set how many generations this simulation will run for.
-- Default number of generations is 1000.
-
-```-h ```
-
-- Displays this guide in program.
-
-```-o crossover_rate```
-
-- Set the crossover rate for the current simulation.
-- Default crossover rate is 2.
-
-```-p population_size```
-
-- Set the population size for the current simulation.
-- Default population size is 10.
-
-```-r```
-
-- Only show percentages of descent from the original genomes.
-
-```-s```
-
-- Degnome selection will occur.
-
-```-u```
-- All degnomes contribute to two offspring.
-
-```-v```
-- Output will be given for every generation.
