--- References_and_Guides/PolygenSim_Commands.md
@@ -1,35 +1,0 @@
-# PolygenSim Commands
-
-```-c chromosome_length```
-
-- Set chromosome length for the current simulation.
-- Default chromosome length is 50.
-
-```-e mutation_effect```
-
-- Set how much a mutation will effect a gene on average.
-- Default mutation effect is 2.
-
-```-g num_generations```
-
-- Set how many generations this simulation will run for.
-- Default number of generations is 1000.
-
-```-h help```
-
-- Displays this guide in program.
-
-```-m mutation_rate```
-
-- Set the mutation rate for the current simulation.
-- Default mutation rate is 1.
-
-```-o crossover_rate```
-
-- Set the crossover rate for the current simulation.
-- Default crossover rate is 2.
-
-```-p population_size```
-
-- Set the population size for the current simulation.
-- Default population size is 100.
