+++ src/ance_degnome.h
