+++ src/degnome.c
