+++ src/degnome.h
