+++ src/fitfunc.c
