+++ src/fitfunc.h
