+++ src/misc.c
