+++ src/xdegnome.c
