+++ src/xfitfunc.c
