+++ src/xmisc.c
